+++ Package.swift
@@ -11,7 +11,8 @@ let package = Package(
     name: "OCCTSwiftViewport",
     platforms: [
         .iOS(.v18),
-        .macOS(.v15)
+        .macOS(.v15),
+        .visionOS(.v1)
     ],
     products: [
         .library(